Fix sample build failure due to intermediate headers being deleted from engine folder. Some modules only have *private* UObject classes which aren't included in the editor zip. UBT detects that there are no longer and UObject classes and deletes the intermediate folder, causing BuildGraph to fail when it detects the change.
#rb Matthew.Griffin
#codereview Matthew.Griffin

[CL 3087148 by Ben Marsh in 4.13 branch]

+++ Engine/Source/Programs/UnrealBuildTool/System/ExternalExecution.cs
@@ -585,10 +585,13 @@ namespace UnrealBuildTool
 				else
 				{
 					// Remove any stale generated code directory
-					if (Module.GeneratedCodeDirectory != null && Module.GeneratedCodeDirectory.Exists())
+					if(!UnrealBuildTool.IsEngineInstalled() || !Module.GeneratedCodeDirectory.IsUnderDirectory(UnrealBuildTool.EngineDirectory))
 					{
-						Log.TraceVerbose("Deleting stale generated code directory: " + Module.GeneratedCodeDirectory.ToString());
-						Directory.Delete(Module.GeneratedCodeDirectory.FullName, true);
+						if (Module.GeneratedCodeDirectory != null && Module.GeneratedCodeDirectory.Exists())
+						{
+							Log.TraceVerbose("Deleting stale generated code directory: " + Module.GeneratedCodeDirectory.ToString());
+							Directory.Delete(Module.GeneratedCodeDirectory.FullName, true);
+						}
 					}
 				}
 			}
